requirements.txt
python-telegram-bot==20.7
pandas
scikit-learn
streamlit
numpy
joblib
httpx